Fix: Converting payment options to a component and calling it in the payment page
